refactor(history): 패키지 구조 개편 및 QueryHistory 변경 반영
