+++ backend/src/main/java/com/gigigenie/domain/history/repository/QueryHistoryRepository.java
@@ -1,14 +1,14 @@
-package com.gigigenie.domain.product.repository;
+package com.gigigenie.domain.history.repository;
 
+import com.gigigenie.domain.history.entity.QueryHistory;
 import com.gigigenie.domain.member.entity.Member;
 import com.gigigenie.domain.product.entity.Product;
-import com.gigigenie.domain.product.entity.QueryHistory;
 import jakarta.validation.constraints.NotNull;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface QueryHistoryRepository extends JpaRepository<QueryHistory, Integer> {
+
     List<QueryHistory> findByMemberAndProduct(@NotNull Member member, @NotNull Product product);
 
     void deleteByMemberAndProduct(@NotNull Member member, @NotNull Product product);
