+++ backend/src/main/java/com/gigigenie/domain/history/service/QueryHistoryServiceImpl.java
@@ -1,15 +1,13 @@
-package com.gigigenie.domain.product.service;
+package com.gigigenie.domain.history.service;
 
+import com.gigigenie.domain.history.dto.QueryHistoryDTO;
+import com.gigigenie.domain.history.entity.QueryHistory;
+import com.gigigenie.domain.history.repository.QueryHistoryRepository;
 import com.gigigenie.domain.member.dto.MemberDTO;
 import com.gigigenie.domain.member.entity.Member;
 import com.gigigenie.domain.member.repository.MemberRepository;
-import com.gigigenie.domain.product.dto.HistoryRequest;
-import com.gigigenie.domain.product.dto.QueryHistoryDTO;
 import com.gigigenie.domain.product.entity.Product;
-import com.gigigenie.domain.product.entity.QueryHistory;
 import com.gigigenie.domain.product.repository.ProductRepository;
-import com.gigigenie.domain.product.repository.QueryHistoryRepository;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -27,40 +25,11 @@ public class QueryHistoryServiceImpl implements QueryHistoryService {
     private final MemberRepository memberRepository;
 
     @Override
-    public void save(HistoryRequest request, Authentication authentication) {
-        Member member = findMember(authentication);
-        Product product = findProduct(request.getProductId());
-
-        List<QueryHistory> histories = new ArrayList<>();
-
-        request.getHistory().forEach(history -> {
-            QueryHistory queryHistory = QueryHistory.builder()
-                .product(product)
-                .member(member)
-                .queryText(history.get("queryText").toString())
-                .responseText(history.get("responseText").toString())
-                .queryTime(((Number) history.get("queryTime")).longValue())
-                .build();
-
-            histories.add(queryHistory);
-        });
-
-        historyRepository.saveAll(histories);
-    }
-
-    @Override
     public List<QueryHistoryDTO> getHistories(Long productId, Authentication authentication) {
         Member member = findMember(authentication);
         Product product = findProduct(productId);
         List<QueryHistory> histories = historyRepository.findByMemberAndProduct(member, product);
         return histories.stream().map(this::entityToDTO).collect(Collectors.toList());
-    }
-
-    @Override
-    public void deleteByMemberAndProduct(Long productId, Authentication authentication) {
-        Member member = findMember(authentication);
-        Product product = findProduct(productId);
-        historyRepository.deleteByMemberAndProduct(member, product);
     }
 
     @Override
